recipes-bsp/u-boot: increase script load addres
Booting with a large a FIT image is failing with the following error:

   Error: ramdisk overwritten
   Ramdisk image is corrupt or invalid

This is because the boot script uses scriptaddr 0x8e000000 to load
the FIT image. However, this may overlap with the fitimage ramdisk
if the ramdisk is big (> 64 MB). Add additional 32M offset to the
script load address by overwriting the boot command and set the
environment to:

setenv scriptaddr 0x90000000

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>